fix: move jdtls back to ftplugin
