+++ modules/nvim/default.nix
@@ -29,6 +29,7 @@ in
     xdg.configFile = {
       "nvim/init.lua".source = ./init.lua;
       "nvim/lua".source = ./lua;
+      "nvim/ftplugin".source = ./ftplugin;
     };
   };
 }
